[Security Solution] Optimize CODEOWNERS: Move Rule Preview related code to Alerts area domain (as part of ##138795)
