+++ x-pack/plugins/security_solution/public/detections/components/rules/rule_preview/use_preview_invocation_count.ts
